backend/requirements.txt
Flask
Flask-Cors
pydantic
supabase
argon2-cffi
Flask-JWT-Extended
python-dotenv
python-dateutil
requests
urllib3
reportlab
